Require SortDirection usage
This removes the need for validation layers.
