+++ src/Cache/Persister/Entity/AbstractEntityPersister.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Doctrine\ORM\Cache\Persister\Entity;
 
 use Doctrine\Common\Collections\Criteria;
-use Doctrine\Common\Collections\Order;
 use Doctrine\DBAL\LockMode;
 use Doctrine\ORM\Cache;
 use Doctrine\ORM\Cache\CollectionCacheKey;
@@ -13,7 +12,6 @@ use Doctrine\ORM\Cache\EntityCacheKey;
 use Doctrine\ORM\Cache\EntityHydrator;
 use Doctrine\ORM\Cache\Logging\CacheLogger;
 use Doctrine\ORM\Cache\Persister\CachedPersister;
-use Doctrine\ORM\Cache\Persister\CompatOrderings;
 use Doctrine\ORM\Cache\QueryCacheKey;
 use Doctrine\ORM\Cache\Region;
 use Doctrine\ORM\Cache\TimestampCacheKey;
@@ -36,8 +34,6 @@ use function sha1;
 
 abstract class AbstractEntityPersister implements CachedEntityPersister
 {
-    use CompatOrderings;
-
     protected UnitOfWork $uow;
     protected ClassMetadataFactory $metadataFactory;
 
@@ -219,13 +215,12 @@ abstract class AbstractEntityPersister implements CachedEntityPersister
     /**
      * Generates a string of currently query
      *
-     * @param string[]|Criteria                                      $criteria
-     * @param array<string, SortDirection>|array<string, Order>|null $orderBy
+     * @param string[]|Criteria                 $criteria
+     * @param array<string, SortDirection>|null $orderBy
      */
     protected function getHash(
         string $query,
         array|Criteria $criteria,
-        // @phpstan-ignore parameter.deprecatedEnum
         array|null $orderBy = null,
         int|null $limit = null,
         int|null $offset = null,
@@ -457,7 +452,7 @@ abstract class AbstractEntityPersister implements CachedEntityPersister
     #[Override]
     public function loadCriteria(Criteria $criteria): array
     {
-        $orderBy     = $this->getOrderings($criteria);
+        $orderBy     = $criteria->getOrderings();
         $limit       = $criteria->getMaxResults();
         $offset      = $criteria->getFirstResult();
         $query       = $this->persister->getSelectSQL($criteria);
