+++ src/QueryBuilder.php
@@ -9,7 +9,6 @@ use Doctrine\Common\Collections\Criteria;
 use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\Deprecations\Deprecation;
-use Doctrine\ORM\Cache\Persister\CompatOrderings;
 use Doctrine\ORM\Internal\NoUnknownNamedArguments;
 use Doctrine\ORM\Query\Expr;
 use Doctrine\ORM\Query\Parameter;
@@ -24,8 +23,6 @@ use function array_keys;
 use function array_unshift;
 use function assert;
 use function count;
-use function func_num_args;
-use function get_debug_type;
 use function implode;
 use function in_array;
 use function is_array;
@@ -38,7 +35,6 @@ use function sprintf;
 use function str_starts_with;
 use function strpos;
 use function strrpos;
-use function strtolower;
 use function substr;
 
 /**
@@ -47,7 +43,6 @@ use function substr;
  */
 class QueryBuilder implements Stringable
 {
-    use CompatOrderings;
     use NoUnknownNamedArguments;
 
     /**
@@ -1225,16 +1220,9 @@ class QueryBuilder implements Stringable
      *
      * @return $this
      */
-    public function orderBy(string|Expr\OrderBy $sort, SortDirection|string|null $order = null): static
+    public function orderBy(string|Expr\OrderBy $sort, SortDirection $order = SortDirection::Ascending): static
     {
-        if (func_num_args() === 1 && is_string($sort)) {
-            $order = SortDirection::Ascending;
-        }
-
-        $orderBy = $sort instanceof Expr\OrderBy ? $sort : new Expr\OrderBy(
-            $sort,
-            $this->getSortDirection($order, __METHOD__),
-        );
+        $orderBy = $sort instanceof Expr\OrderBy ? $sort : new Expr\OrderBy($sort, $order);
 
         return $this->add('orderBy', $orderBy);
     }
@@ -1244,15 +1232,11 @@ class QueryBuilder implements Stringable
      *
      * @return $this
      */
-    public function addOrderBy(string|Expr\OrderBy $sort, SortDirection|string|null $order = null): static
+    public function addOrderBy(string|Expr\OrderBy $sort, SortDirection $order = SortDirection::Ascending): static
     {
-        if (func_num_args() === 1 && is_string($sort)) {
-            $order = SortDirection::Ascending;
-        }
-
         $orderBy = $sort instanceof Expr\OrderBy ? $sort : new Expr\OrderBy(
             $sort,
-            $this->getSortDirection($order, __METHOD__),
+            $order,
         );
 
         return $this->add('orderBy', $orderBy, true);
@@ -1286,7 +1270,7 @@ class QueryBuilder implements Stringable
             }
         }
 
-        $orderings = $this->getOrderings($criteria);
+        $orderings = $criteria->getOrderings();
 
         foreach ($orderings as $sort => $order) {
             $hasValidAlias = false;
@@ -1301,8 +1285,7 @@ class QueryBuilder implements Stringable
                 $sort = $allAliases[0] . '.' . $sort;
             }
 
-            /** @phpstan-ignore property.deprecatedEnum */
-            $this->addOrderBy($sort, $order instanceof SortDirection ? $order : $order->value);
+            $this->addOrderBy($sort, $order);
         }
 
         // Overwrite limits only if they was set in criteria
@@ -1509,30 +1492,5 @@ class QueryBuilder implements Stringable
         }
 
         $this->parameters = new ArrayCollection($parameters);
-    }
-
-    private function getSortDirection(SortDirection|string|null $order, string $method): SortDirection
-    {
-        if ($order instanceof SortDirection) {
-            return $order;
-        }
-
-        Deprecation::trigger(
-            'doctrine/orm',
-            'https://github.com/doctrine/orm/issues/11313',
-            'Passing %s as $order to %s() is deprecated, use an instance of SortDirection instead.',
-            get_debug_type($order),
-            $method,
-        );
-
-        if ($order === null) {
-            return SortDirection::Ascending;
-        }
-
-        return match (strtolower($order)) {
-            'asc' => SortDirection::Ascending,
-            'desc' => SortDirection::Descending,
-            default => throw new InvalidArgumentException(sprintf('Invalid sort direction "%s".', $order)),
-        };
     }
 }
